requirements/dev.txt
-r base.txt

atomicwrites==1.2.1
attrs==18.2.0
factory-boy==2.11.1
Faker==1.0.1
more-itertools==4.3.0
pluggy==0.8.0
py==1.7.0
pytest==4.0.2
pytest-django==3.4.4
python-dateutil==2.7.5
six==1.12.0
text-unidecode==1.2
